style: join, userResearch 스타일링 변경
