+++ components/pages/join/Description.js
@@ -1,6 +1,6 @@
 const Description = ({ description }) => {
   return (
-    <div className="multiple-24-b text-neutralgray-900">
+    <div className="headline-s text-neutralgray-900">
       <div>{description[0]}</div>
       <div>{description[1]}</div>
     </div>
