+++ components/pages/userResearch/Job.js
@@ -43,66 +43,69 @@ export default function Job() {
   }, []);
 
   return (
-    <div ref={wrapperRef} className="relative">
-      <Listbox value={selected} onChange={setSelected}>
-        <div className="grid">
-          <Listbox.Button
-            className={`z-10 text-left p-4 w-full text-base font-medium 
-            ${clickArea ? "rounded-t-lg" : "rounded-lg"} 
-            ${selectedCSS} 
-            ${borderCSS}`}
-            onClick={() => setClickArea((prev) => !prev)}
-          >
-            <Listbox.Label
-              className={`flex justify-between items-center cursor-pointer 
-              ${selected === false ? labelCSS : selectedCSS} `}
+    <div>
+      <div>종사산업</div>
+      <div ref={wrapperRef} className="relative">
+        <Listbox value={selected} onChange={setSelected}>
+          <div className="grid">
+            <Listbox.Button
+              className={`z-10 text-left p-4 w-full text-base font-medium
+              ${clickArea ? "rounded-t-lg" : "rounded-lg"}
+              ${selectedCSS}
+              ${borderCSS}`}
+              onClick={() => setClickArea((prev) => !prev)}
             >
-              {!selected ? "산업군 선택" : selected.name}
-              <ArrowIcon
-                className={`${
-                  clickArea ? "rotate-180" : ""
-                } transition-transform duration-500`}
-              />
-            </Listbox.Label>
-          </Listbox.Button>
-          <Transition
-            as={Listbox.Options}
-            show={clickArea}
-            enter="transition ease-out duration-200"
-            enterFrom="transform -translate-y-10 opacity-0"
-            enterTo="transform translate-y-0 opacity-100"
-            leave="transition ease-in duration-200"
-            leaveFrom="transform translate-y-0 opacity-100"
-            leaveTo="transform -translate-y-10 opacity-0"
-          >
-            {clickArea && (
-              <Listbox.Options
-                id="onboardIndustryBox"
-                className={`z-0 overflow-scroll cursor-pointer rounded-b-lg ${height} border-x-2 border-b-2 border-purple-200`}
+              <Listbox.Label
+                className={`flex justify-between items-center cursor-pointer
+                ${selected === false ? labelCSS : selectedCSS} `}
               >
-                {industries?.map((ind) => {
-                  <Listbox.Option key={ind.id} value={ind} as={Fragment}>
-                    {({ active }) => {
-                      return (
-                        <li
-                          className={`${
-                            active
-                              ? "text-neutralgray-900"
-                              : "bg-white text-neutralgray-900"
-                          } flex justify-between p-4 single-16-m font-medium transition-colors duration-120 hover:bg-purple-50 active:bg-purple-50`}
-                          onClick={() => setClickArea(false)}
-                        >
-                          {ind.name}
-                        </li>
-                      );
-                    }}
-                  </Listbox.Option>;
-                })}
-              </Listbox.Options>
-            )}
-          </Transition>
-        </div>
-      </Listbox>
+                {!selected ? "산업군 선택" : selected.name}
+                <ArrowIcon
+                  className={`${
+                    clickArea ? "rotate-180" : ""
+                  } transition-transform duration-500`}
+                />
+              </Listbox.Label>
+            </Listbox.Button>
+            <Transition
+              as={Listbox.Options}
+              show={clickArea}
+              enter="transition ease-out duration-200"
+              enterFrom="transform -translate-y-10 opacity-0"
+              enterTo="transform translate-y-0 opacity-100"
+              leave="transition ease-in duration-200"
+              leaveFrom="transform translate-y-0 opacity-100"
+              leaveTo="transform -translate-y-10 opacity-0"
+            >
+              {clickArea && (
+                <Listbox.Options
+                  id="onboardIndustryBox"
+                  className={`z-0 overflow-scroll cursor-pointer rounded-b-lg ${height} border-x-2 border-b-2 border-purple-200`}
+                >
+                  {industries?.map((ind) => {
+                    <Listbox.Option key={ind.id} value={ind} as={Fragment}>
+                      {({ active }) => {
+                        return (
+                          <li
+                            className={`${
+                              active
+                                ? "text-neutralgray-900"
+                                : "bg-white text-neutralgray-900"
+                            } flex justify-between p-4 single-16-m font-medium transition-colors duration-120 hover:bg-purple-50 active:bg-purple-50`}
+                            onClick={() => setClickArea(false)}
+                          >
+                            {ind.name}
+                          </li>
+                        );
+                      }}
+                    </Listbox.Option>;
+                  })}
+                </Listbox.Options>
+              )}
+            </Transition>
+          </div>
+        </Listbox>
+      </div>
     </div>
   );
 }
