+++ constants/user_research_pages.tsx
@@ -2,37 +2,25 @@ export const pages = [
   0,
   {
     id: 1,
-    header_1_1: "가입완료!😎",
-    header_1_2: "뉴스레터 추천을 위한",
-    header_1_3: "구독 이메일이 생성되었어요!",
+    header_1_1: "뉴스레터 구독을 위한",
+    header_1_2: "이메일이 생성되었어요.",
     caption: "이제 뉴독의 구독 이메일로 뉴스레터 구독을 신청하고",
     caption_2: "아티클을 수신받아 간편하게 관리할 수 있어요.",
   },
   {
     id: 2,
-    header_1_1: "반가워요👋",
-    header_1_2: `뉴스레터 추천을 위해`,
-    header_1_3: `님에 대해 더 알려주세요.`,
-    caption: "종사 산업과 관심사를 선택하여",
+    header_1_1: "뉴스레터 추천을 위해",
+    header_1_2: `님에 대해`,
+    header_1_3: `더 알려주세요.`,
+    caption: "종사 산업과 관심사를 선택하면",
     caption_2: "내게 도움이 될 뉴스레터를 만나볼 수 있어요!",
   },
+
   {
     id: 3,
-    header_1_1: "종사 중인 산업을",
-    header_1_2: `선택해 주세요.`,
-    caption: "선택하신 산업과 관련된 뉴스레터를 찾아드려요.",
-  },
-  {
-    id: 4,
-    header_1_1: "더 정확한 추천을 위해 ",
-    header_1_2: "관심사를 선택해 주세요.",
-    caption: "최소 3가지 이상을 선택해주세요.",
-  },
-  {
-    id: 5,
     header_1_1: "님을 위한 ",
     header_1_2: "맞춤형 뉴스레터가 도착했어요!",
-    caption: "구독한 뉴스레터는 발행일에 맞춰 홈으로 배달해드려요.",
+    caption: "구독한 뉴스레터는 발행일에 맞춰 배달해드려요.",
     caption_2: "[구독하기]를 누르면 구독 이메일이 자동으로 복사돼요.",
   },
 ];
